+++ jni/ia/NeuralNetwork.h
@@ -8,17 +8,17 @@
 class NeuralNetwork {
     std::vector<Layer> layers;
     unsigned int epochs,max_epochs;
-    double threshold;
+    float threshold;
 
     unsigned int inputSize, outputSize;
 
     public:
     NeuralNetwork(unsigned int inputSize, unsigned int outputSize);
-    void addLayer(unsigned int outputSize,ACTIVATION_TYPE activationType,double learningRate);
-    void addLayer(Matrix w,ACTIVATION_TYPE activationType,double learningRate);
+    void addLayer(unsigned int outputSize,ACTIVATION_TYPE activationType,float learningRate);
+    void addLayer(Matrix w,ACTIVATION_TYPE activationType,float learningRate);
     void setEpochs(unsigned int);
     unsigned int getMaxEpochs();
-    void setThreshold(double th);
+    void setThreshold(float th);
     void train(Vector input,Vector output);
     void results();
     Vector ia(Vector v);
